+++ core/impl/src/main/java/com/blazebit/persistence/impl/function/stringjsonagg/OracleStringJsonAggFunction.java
@@ -5,42 +5,35 @@
 
 package com.blazebit.persistence.impl.function.stringjsonagg;
 
-import com.blazebit.persistence.impl.function.chr.ChrFunction;
-import com.blazebit.persistence.impl.function.concat.ConcatFunction;
-import com.blazebit.persistence.impl.function.groupconcat.AbstractGroupConcatFunction;
-import com.blazebit.persistence.impl.function.replace.ReplaceFunction;
 import com.blazebit.persistence.spi.FunctionRenderContext;
 
 /**
- * @author Christian Beikov
- * @since 1.5.0
+ * Oracle Database 12c Release 2 included support for JSON_OBJECT and JSON_ARRAYAGG
+ *
+ * @author Michael Saull
+ * @since 1.6.x?
  */
-public class OracleStringJsonAggFunction extends GroupConcatBasedStringJsonAggFunction {
-
-    public OracleStringJsonAggFunction(AbstractGroupConcatFunction groupConcatFunction, ChrFunction chrFunction, ReplaceFunction replaceFunction, ConcatFunction concatFunction) {
-        super(groupConcatFunction, chrFunction, replaceFunction, concatFunction);
-    }
+public class OracleStringJsonAggFunction extends AbstractStringJsonAggFunction {
 
     @Override
     public void render(FunctionRenderContext context) {
-        if ((context.getArgumentsSize() & 1) == 1) {
+        int argumentsCount = context.getArgumentsSize();
+        if ((argumentsCount & 1) == 1) {
             throw new RuntimeException("The string_json_agg function needs an even amount of arguments <key1>, <value1>, ..., <keyN>, <valueN>! args=" + context);
         }
-        context.addChunk("('[' || dbms_xmlgen.convert(substr(xmlagg(xmlelement(e,to_clob(',')||");
 
-        StringBuilder sb = new StringBuilder();
-        render(sb, context);
-        context.addChunk(sb.toString());
-        context.addChunk(").extract('//text()')).getClobVal(),2),1) || ']')");
+        context.addChunk("json_arrayagg(json_object(");
+        writeArgumentPair(context, 0);
+        for (int i = 2; i < argumentsCount; i += 2) {
+            context.addChunk(",");
+            writeArgumentPair(context, i);
+        }
+        context.addChunk("))");
     }
 
-    @Override
-    protected String coalesceStart() {
-        return "coalesce(nullif(";
-    }
-
-    @Override
-    protected String coalesceEnd() {
-        return ",'\"\"'),'null')";
+    private static void writeArgumentPair(FunctionRenderContext context, int i) {
+        context.addArgument(i);
+        context.addChunk(":");
+        context.addArgument(i + 1);
     }
 }
