+++ core/impl/src/main/java/com/blazebit/persistence/impl/function/tostringjson/OracleToStringJsonFunction.java
@@ -5,62 +5,45 @@
 
 package com.blazebit.persistence.impl.function.tostringjson;
 
-import com.blazebit.persistence.impl.function.chr.ChrFunction;
-import com.blazebit.persistence.impl.function.concat.ConcatFunction;
-import com.blazebit.persistence.impl.function.groupconcat.AbstractGroupConcatFunction;
-import com.blazebit.persistence.impl.function.replace.ReplaceFunction;
 import com.blazebit.persistence.impl.util.SqlUtils;
 import com.blazebit.persistence.spi.FunctionRenderContext;
-import com.blazebit.persistence.spi.LateralStyle;
 
 /**
- * @author Christian Beikov
- * @since 1.5.0
+ * Adapted from MySQLToStringJsonFunction by
+ * Christian Beikov
+ *
+ * @author Michael Saull
+ * @since 1.6.x?
  */
-public class OracleToStringJsonFunction extends GroupConcatBasedToStringJsonFunction {
+public class OracleToStringJsonFunction extends AbstractToStringJsonFunction {
 
-    private static final String ELEMENT_POST_CHUNK = ").extract('//text()'))";
-    private static final String AGGREGATE_POST_CHUNK = ".getClobVal(),2),1)) || ']')";
+    private static final String START_CHUNK = "(select json_arrayagg(json_object('";
+    private static final String ELEMENT_POST_CHUNK = ")";
+    private static final String AGGREGATE_POST_CHUNK = "RETURNING CLOB)"; //To avoid VARCHAR2(4000) limit
     private static final String POST_CHUNK = ELEMENT_POST_CHUNK + AGGREGATE_POST_CHUNK;
-
-    public OracleToStringJsonFunction(AbstractGroupConcatFunction groupConcatFunction, ChrFunction chrFunction, ReplaceFunction replaceFunction, ConcatFunction concatFunction) {
-        super("(select ('[' || (dbms_xmlgen.convert(substr(xmlagg(xmlelement(e,to_clob(',')||", POST_CHUNK, false, groupConcatFunction, chrFunction, replaceFunction, concatFunction, LateralStyle.LATERAL);
-    }
 
     @Override
     public void render(FunctionRenderContext context, String[] fields, String[] selectItemExpressions, String subquery, int fromIndex) {
+        context.addChunk(START_CHUNK);
+
         int orderByIndex = SqlUtils.indexOfOrderBy(subquery, fromIndex);
         if (orderByIndex == -1) {
-            context.addChunk(preChunk);
-            StringBuilder sb = new StringBuilder(fromIndex);
-            render(sb, fields, selectItemExpressions);
-            context.addChunk(sb.toString());
-            context.addChunk(postChunk);
+            renderJsonObjectArguments(context, fields, selectItemExpressions);
+            context.addChunk(POST_CHUNK);
             context.addChunk(subquery.substring(fromIndex));
         } else {
-            if (SqlUtils.indexOfLimit(subquery, orderByIndex) == -1 && SqlUtils.indexOfFetchFirst(subquery, orderByIndex) == -1) {
-                context.addChunk(preChunk);
-
-                StringBuilder sb = new StringBuilder(fromIndex);
-                render(sb, fields, selectItemExpressions);
-                context.addChunk(sb.toString());
-
-                context.addChunk(ELEMENT_POST_CHUNK);
+            int limitIndex = SqlUtils.indexOfLimit(subquery, orderByIndex);
+            if (limitIndex == -1) {
+                renderJsonObjectArguments(context, fields, selectItemExpressions);
+                context.addChunk(POST_CHUNK);
                 context.addChunk(" OVER (");
                 context.addChunk(subquery.substring(orderByIndex));
                 context.addChunk(")");
-                context.addChunk(AGGREGATE_POST_CHUNK);
                 context.addChunk(subquery.substring(fromIndex, orderByIndex));
-                context.addChunk(")");
             } else {
-                context.addChunk(preChunk);
-
-                StringBuilder sb = new StringBuilder(fromIndex);
-                render(sb, fields, selectItemExpressions);
-                context.addChunk(sb.toString());
-
-                context.addChunk(postChunk);
-                context.addChunk(" from lateral(select ");
+                renderJsonObjectArguments(context, fields, fields); //I am not sure if this part is correct limit syntax for oracle? Might need to use the lateral join from OracleToStringJsonFunction (V1)
+                context.addChunk(POST_CHUNK);
+                context.addChunk(" from (select ");
                 for (int i = 0; i < fields.length; i++) {
                     if (i != 0) {
                         context.addChunk(",");
@@ -75,13 +58,18 @@ public class OracleToStringJsonFunction extends GroupConcatBasedToStringJsonFunc
         }
     }
 
-    @Override
-    protected String coalesceStart() {
-        return "coalesce(nullif(";
+    private void renderJsonObjectArguments(FunctionRenderContext context, String[] fields, String[] selectItemExpressions) {
+        writeFieldItemPair(context, fields[0], selectItemExpressions[0]);
+        for (int i = 1; i < fields.length; i++) { //abstract ensures size
+            context.addChunk(",'");
+            writeFieldItemPair(context, fields[i], selectItemExpressions[i]);
+        }
     }
 
-    @Override
-    protected String coalesceEnd(String field) {
-        return ",'\"\"'),'null')";
+    private void writeFieldItemPair(FunctionRenderContext context, String field, String itemExpression) {
+        context.addChunk(field);
+        context.addChunk("':");
+        context.addChunk(itemExpression); //Do we need to do a cast to string type like mysql does?
     }
+
 }
